Add packager.io security warning and Debian 9 (#232)
The latest version on packager.io is 0.11.34 (released Nov 22, 2017, built January 2019). The latest gogs release, which includes security fixes, is 0.11.86 (January 2019).
The "at your own risk" warning is good, but I don't think it's sufficient at this point - continuing to refer users to packager.io without an explicit warning when those builds are this outdated would be reckless. Removing the links would also be fine.
If anyone wants to update it, it looks like it lives at https://github.com/pkgr/gogs